broke up out of box Create React App into exmpaple file structure to get the pattern started
