+++ ui/src/components/ReactExampleHeader/ReactExampleHeader.css
@@ -1,7 +1,3 @@
-.App {
-  text-align: center;
-}
-
 .App-logo {
   height: 40vmin;
   pointer-events: none;
