+++ ui/src/components/ReactExampleHeader/index.js
@@ -1,0 +1,3 @@
+import ReactExampleHeader from './ReactExampleHeader'
+
+export default ReactExampleHeader;
