+++ ui/src/containers/App/App.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import ReactEmapleHeader from '../../components/ReactExampleHeader';
+import '../../styles/index.css';
+
+function App() {
+  return (
+    <div className="App">
+      <ReactEmapleHeader />
+    </div>
+  );
+}
+
+export default App;
